ユーザー論理削除：復帰ボタン実装

+++ app/controllers/users_controller.rb
@@ -46,15 +46,20 @@ class UsersController < ApplicationController
 	end
 
 	def destroy
-		user = User.find(params[:id])
-		if  user.destroy
-			if admin_signed_in?
-				redirect_to admins_path
+		user = User.with_deleted.find(params[:id])
+		unless user.deleted_at.present?
+			if  user.destroy
+				if admin_signed_in?
+					redirect_to admins_path
+				else
+					redirect_to books_path
+				end
 			else
-				redirect_to books_path
+				redirect_to edit_book_path
 			end
 		else
-			redirect_to edit_book_path
+			user.restore
+			redirect_to admins_path
 		end
 	end
